refactor to reduce complexity of datasource
